fix: Update version and fix data saving function
This commit updates the package version from 2.0.1-beta to 2.0.2-beta and fixes an issue with the function of saving or not saving data in the database. It modifies the package.json and CHANGELOG.md files.

Note: The CHANGELOG.md file includes a detailed list of changes and can be reviewed for more information.